requirements.txt
beautifulsoup4>=4.13.0
lxml>=4.9.3
python-dotenv>=1.0.0
requests>=2.32.0
selenium>=4.18.0
